requirements.txt
mysql-connector-python>=8.0.33
PyYAML>=6.0
pandas>=2.0.0
google-cloud-storage>=2.10.0
pyarrow>=12.0.0
